requirements.txt
asgiref==3.4.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
dataclasses==0.8
defusedxml==0.7.1
Django==3.0.6
django-allauth==0.51.0
django-autoslug==1.9.8
django-crispy-forms==1.13.0
django-markdownx==3.0.1
docopt==0.6.2
ExifRead==3.0.0
idna==3.3
imageio==2.15.0
importlib-metadata==4.8.3
Markdown==3.3.7
numpy==1.19.5
oauthlib==3.2.0
opencv-contrib-python==4.5.5.64
packbits==0.6
Pillow==8.4.0
psd-tools3==1.8.2
PyBundle==1.0.6
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
torch==1.10.2
torchvision==0.11.3
typing-extensions==4.1.1
urllib3==1.26.9
whitenoise==5.3.0
zipp==3.6.0
